% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{addPmToolbar}
\alias{addPmToolbar}
\title{Add 'leaflet-pm' Toolbar to a Leaflet 'htmlwidget'}
\usage{
addPmToolbar(map = NULL, targetLayerId = NULL, targetGroup = NULL,
  toolbarOptions = pmToolbarOptions(), drawOptions = pmDrawOptions(),
  editOptions = pmEditOptions(), cutOptions = pmCutOptions())
}
\arguments{
\item{map}{\code{leaflet} map htmlwidget to which the toolbar should be added}

\item{targetLayerId}{\code{string} id of the feature to edit}

\item{targetGroup}{\code{string} name of the group to edit}

\item{toolbarOptions}{\code{\link{pmToolbarOptions}}}

\item{drawOptions}{\code{\link{pmDrawOptions}}}

\item{editOptions}{\code{\link{pmEditOptions}}}

\item{cutOptions}{\code{\link{pmCutOptions}}}
}
\value{
\code{leaflet} htmlwidget
}
\description{
Add 'leaflet-pm' Toolbar to a Leaflet 'htmlwidget'
}
\examples{
\dontrun{
  library(leaflet)
  library(leafpm)

  # quick example
  leaflet() \%>\%
    addTiles() \%>\%
    addPmToolbar()

  # customizing with options
  leaflet() \%>\%
    addTiles() \%>\%
    addPmToolbar(
      toolbarOptions = pmToolbarOptions(drawMarker = FALSE, position = "topright"),
      drawOptions = pmDrawOptions(snappable = FALSE, allowSelfIntersection = FALSE)
      #editOptions = pmEditOptions(...)
      #cutOptions = pmCutOptions(...)
    )
}
}
