% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pmCutOptions}
\alias{pmCutOptions}
\title{Set Options for 'leaflet.pm' Cut Mode}
\usage{
pmCutOptions(snappable = FALSE, cursorMarker = FALSE)
}
\arguments{
\item{snappable}{\code{logical} to snap while drawing. Default is \code{TRUE}.}

\item{cursorMarker}{\code{logical} to show a marker at the cursor.  Default is \code{TRUE}.}
}
\value{
\code{list} of options
}
\description{
Set Options for 'leaflet.pm' Cut Mode
}
