% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pmDrawOptions}
\alias{pmDrawOptions}
\title{Title}
\usage{
pmDrawOptions(snappable = TRUE, snapDistance = 20, tooltips = TRUE,
  cursorMarker = TRUE, finishOnDoubleClick = FALSE, finishOn = NULL,
  allowSelfIntersection = TRUE, templineStyle = list(),
  hintlineStyle = list(color = "#3388ff", dashArray = "5,5"),
  markerStyle = list(draggable = TRUE))
}
\arguments{
\item{markerStyle}{}
}
\description{
Title
}
