% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{removePmToolbar}
\alias{removePmToolbar}
\title{Removes the 'leaflet.pm' toolbar}
\usage{
removePmToolbar(map, clearFeatures = FALSE)
}
\arguments{
\item{map}{\code{leaflet} map htmlwidget from which the toolbar should be removed}

\item{clearFeatures}{whether to clear the map of drawn features. currently not working}
}
\value{
\code{leaflet} htmlwidget
}
\description{
Removes the 'leaflet.pm' toolbar
}
