% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{addPmToolbar}
\alias{addPmToolbar}
\title{Add 'leaflet-pm' Toolbar to a Leaflet 'htmlwidget'}
\usage{
addPmToolbar(map = NULL, targetLayerId = NULL, targetGroup = NULL,
  toolbarOptions = pmToolbarOptions(), drawOptions = pmDrawOptions(),
  editOptions = pmEditOptions())
}
\arguments{
\item{editOptions}{}
}
\value{
leaflet htmlwidget
}
\description{
Add 'leaflet-pm' Toolbar to a Leaflet 'htmlwidget'
}
