% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pmEditOptions}
\alias{pmEditOptions}
\title{Title}
\usage{
pmEditOptions(snappable = TRUE, snapDistance = 20,
  allowSelfIntersection = TRUE, draggable = TRUE)
}
\arguments{
\item{draggable}{}
}
\description{
Title
}
