% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pmToolbarOptions}
\alias{pmToolbarOptions}
\title{Title}
\usage{
pmToolbarOptions(drawMarker = TRUE, drawPolygon = TRUE,
  drawPolyline = TRUE, drawCircle = TRUE, drawRectangle = TRUE,
  editMode = TRUE, cutPolygon = TRUE, dragPolygon = FALSE,
  removalMode = TRUE, position = "topleft")
}
\arguments{
\item{position}{}
}
\description{
Title
}
